Forgot this in last commit
git-svn-id: http://solemnwarning.net/svn/octradius/trunk@208 8eaceb45-6f80-42b0-b3a5-e78a358d2bcd

+++ src/tile_anims.cpp
@@ -23,7 +23,7 @@ namespace TileAnimators {
 				int ry = (t->screen_y + t->height * TILE_HEIGHT_FACTOR) - (center->screen_y + center->height * TILE_HEIGHT_FACTOR);
 				t->anim_height = t->height;
 				t->animating = true;
-				t->anim_delay = sqrt(pow(rx, 2) + pow(ry, 2)) * delay_factor;
+				t->anim_delay = sqrt(pow((double)rx, 2) + pow((double)ry, 2)) * delay_factor;
 				t->initial_elevation = t->height;
 				if (mode == ABSOLUTE)
 					t->final_elevation = target_elevation;
